feat: move TorServiceStatus, StringFeedState, BookmarkType to commons
Move sealed classes and enums to commons with explicit import updates:
- TorServiceStatus: sealed class for Tor connection states
- StringFeedState: sealed class for string feed loading states
- BookmarkType: enum for bookmark categories

Consumers updated with explicit imports from commons package to work
around Kotlin typealias limitation with nested type resolution.
Part of KMP iOS migration tracked in #2238.
